Video 3/drf_site/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
